docs: refine the example for segmentation problem type (#124)
