+++ .gitignore
@@ -169,5 +169,5 @@ node_modules/
 .python-version
 .ruff_cache
 .coverage
-examples/lime/cloudsen12_models/UNetMobV2_V1.pt
+examples/segmentation/cloudsen12_models/UNetMobV2_V1.pt
 .vscode/settings.json
