+++ .pre-commit-config.yaml
@@ -59,7 +59,7 @@ repos:
       - id: debug-statements
       - id: requirements-txt-fixer
       - id: check-added-large-files
-        args: ["--maxkb=5000"]
+        args: ["--maxkb=9000"]
       - id: check-ast
       - id: check-yaml
       - id: check-toml
