--- examples/lime/cloudsen12_models/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "0.4.3"
