+++ examples/segmentation/cloudsen12_models/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.4.3"
